Made gcWeight a double (used to be a float)

+++ castor/io/StreamDiskCopyCnv.cpp
@@ -114,7 +114,7 @@ castor::IObject* castor::io::StreamDiskCopyCnv::createObj(castor::IAddress* addr
   std::string path;
   ad->stream() >> path;
   object->setPath(path);
-  float gcWeight;
+  double gcWeight;
   ad->stream() >> gcWeight;
   object->setGcWeight(gcWeight);
   u_signed64 creationTime;
